refactor: adjust task view to current project architecture
